Finalized local project structure
A finalized project structure that will suit a local/dev
environment. If/when the airflow instance is to be hosted the
project style of the pipeline repos may need to be changed.
Name of the Github repo and some folder structure names will likely
be tweaked, but overall structure should remain intact.

The idea PROD structure would be 1 airflow instance hosted remotely
that can load  multiple pipelines, which would each have their own
git repo.