Fully implemented transformation task
A DatabricksSubmitRunOperator task loads data from the
S3 unstructured bucket and stages and models the data in
delta tables on the Databricks cluster. A subsequent task
to load the data into the data warehouse should be
implemented. New keys "spark_cluster_id" and
"spark_notebook_path" added to properties_batch_config
Variable.

The notebook(s) associated with this task have also been
added to this repo and can be managed by Databricks by
configuring Git integration using Personal Access Tokens.
It should be noted that Community Editions of Databricks
will not work with this task as as the Community Edition
doesn’t allow for API access, which is how the operator
triggers the task.